digit classififer/requirements.txt
flask
numpy
tensorflow
flask_cors
pillow